requirements.txt
langchain-openai
llama-index
streamlit
dspy-ai
qdrant-client
tavily-python
python-dotenv
matplotlib
numpy
pandas
openai
requests
transformers